+++ common/RELEASE_NOTES.md
@@ -2,6 +2,8 @@
 
 ## in develop
 
+* Fix `JsUtils.makeDeterministic` to handle JsArrays sorting
+
 ## 0.11.2 (2022-02-25)
 
 * Minor changes to JSON formatting